Create wip prefix_urls table
